requirements.txt
pymysql~=0.9.3
flask~=1.1.2
python-dotenv~=0.14.0